+++ aimiaobi-20230801/alibabacloud_aimiaobi20230801/models/_hottopic_news.py
@@ -1,0 +1,95 @@
+# -*- coding: utf-8 -*-
+# This file is auto-generated, don't edit it. Thanks.
+from __future__ import annotations
+
+from typing import List
+
+from alibabacloud_aimiaobi20230801 import models as main_models
+from darabonba.model import DaraModel
+
+class HottopicNews(DaraModel):
+    def __init__(
+        self,
+        comments: List[main_models.HottopicNewsComments] = None,
+        content: str = None,
+        title: str = None,
+        url: str = None,
+    ):
+        self.comments = comments
+        self.content = content
+        self.title = title
+        self.url = url
+
+    def validate(self):
+        if self.comments:
+            for v1 in self.comments:
+                 if v1:
+                    v1.validate()
+
+    def to_map(self):
+        result = dict()
+        _map = super().to_map()
+        if _map is not None:
+            result = _map
+        result['Comments'] = []
+        if self.comments is not None:
+            for k1 in self.comments:
+                result['Comments'].append(k1.to_map() if k1 else None)
+
+        if self.content is not None:
+            result['Content'] = self.content
+
+        if self.title is not None:
+            result['Title'] = self.title
+
+        if self.url is not None:
+            result['Url'] = self.url
+
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        self.comments = []
+        if m.get('Comments') is not None:
+            for k1 in m.get('Comments'):
+                temp_model = main_models.HottopicNewsComments()
+                self.comments.append(temp_model.from_map(k1))
+
+        if m.get('Content') is not None:
+            self.content = m.get('Content')
+
+        if m.get('Title') is not None:
+            self.title = m.get('Title')
+
+        if m.get('Url') is not None:
+            self.url = m.get('Url')
+
+        return self
+
+class HottopicNewsComments(DaraModel):
+    def __init__(
+        self,
+        text: str = None,
+    ):
+        self.text = text
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        result = dict()
+        _map = super().to_map()
+        if _map is not None:
+            result = _map
+        if self.text is not None:
+            result['Text'] = self.text
+
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('Text') is not None:
+            self.text = m.get('Text')
+
+        return self
+